feat: 초기 폴더 구조 정리 및 styled-components 적용
