+++ frontend/src/App.jsx
@@ -1,0 +1,7 @@
+import { Main } from "./components";
+
+function App() {
+  return <Main />;
+}
+
+export default App;
