+++ frontend/src/components/common/Header.style.jsx
@@ -1,0 +1,6 @@
+import styled from "styled-components";
+
+export const Title = styled.h2`
+  font-size: 72px;
+  color: #ff0000;
+`;
