+++ frontend/src/components/detail/Detail.jsx
@@ -1,0 +1,3 @@
+export function Detail() {
+  return <div>Detail 페이지</div>;
+}
