+++ frontend/src/components/detail/index.js
@@ -1,0 +1,1 @@
+export * from "./Detail";
