+++ frontend/src/components/index.js
@@ -1,0 +1,3 @@
+export * from "./common";
+export * from "./detail";
+export * from "./main";
