+++ frontend/src/components/main/Main.jsx
@@ -1,0 +1,10 @@
+import Header from "../common/Header";
+
+export function Main() {
+  return (
+    <div>
+      <h1>Main 페이지</h1>
+      <Header />
+    </div>
+  );
+}
