--- frontend/src/pages/Detail.css
@@ -1,147 +1,0 @@
-@import url('https://fonts.googleapis.com/css2?family=Do+Hyeon&display=swap');
-
-body {
-    font-family: "Do Hyeon", '굴림';
-    font-size: 40px;
-    color: #000000;
-}
-
-h2 {
-    font-family: "Do Hyeon", "굴림";
-    font-size: 72px;
-    color: #00462A;
-    margin: 0;
-    line-height: 1.5;
-    font-style: normal;
-    font-weight: 400;
-}
-
-h3 {
-    font-family: "Do Hyeon", "굴림";
-    font-size: 56px;
-    color: #000000;
-    margin: 0;
-    line-height: 1.4;
-    font-style: normal;
-    font-weight: 400;
-}
-
-.header {
-    display: flex;
-    align-items: center;
-    justify-content: space-between;
-    position: fixed;
-    z-index: 1000;
-    top: 0;
-    left: 0;
-    width: 100%;
-    height: 128px;
-    background-color: #00462A;
-    padding: 10px 20px;
-    box-sizing: border-box;
-    overflow: visible;
-}
-
-.logo img {
-    height: 76px;
-    flex-shrink: 0;
-    padding-left: 45px;
-}
-
-.nav {
-    display: flex;
-    gap: 29px;
-    align-items: center;
-}
-
-header nav a {
-    text-decoration: none;
-    color: #00462A;
-    font-weight: normal;
-}
-
-.nav-button {
-    display: flex;
-    align-items: center;
-    justify-content: center;
-    color: #00462A;
-    font-family: "Do Hyeon", "굴림";
-    font-size: 56px;
-    font-style: normal;
-    font-weight: 400;
-    line-height: normal;
-    text-align: center;
-    width: 252px;
-    height: 91px;
-    flex-shrink: 0;
-    background: url('../images/button.png') no-repeat center;
-    background-size: contain;
-    border: none;
-    cursor: pointer;
-
-    transform: translateX(-45px);
-}    
-      
-.mypage-button {
-    display: inline-block;
-    width: 91px;
-    height: 91px;
-    flex-shrink: 0;
-    background: url('../images/button_mypage.png') no-repeat center;
-    background-size: cover;
-    cursor: pointer;
-
-    transform: translateX(-45px);
-}
-
-main {
-    margin-top: 128px;
-    margin-left: 60px;
-    margin-right: 60px;
-}
-
-.title {
-    display: flex;
-    align-items: baseline;
-    padding-top: 65px;
-    gap: 35px;
-}
-
-.title-line {
-    border: none;
-    align-items: center;
-    border-top: 5px solid #00462A;
-    margin-top: 10px;
-    width: 100%;
-}
-
-.product-detail {
-    display: flex;
-    flex-direction: column;
-    gap: 20px;
-    margin: 20px;
-}
-
-.product-main {
-    display: flex;
-    gap: 20px;
-}
-
-.product-image img {
-    max-width: 500px;
-    height: auto;
-}
-
-.product-info {
-    display: flex;
-    flex-direction: column;
-    justify-content: flex-start;
-    gap: 10px;
-}
-
-.seller-info {
-    display: flex;
-    flex-direction: column;
-    justify-content: flex-start;
-    margin: 20px;
-}
